added middleware for handling storing auth tokens and user data.

+++ controllers/blog.js
@@ -16,25 +16,11 @@ blogRouter.get('/:id', async (request, response) => {
   blogs ? response.json(blogs) : response.status(404).end();
 });
 
-const getTokenFrom = (request) => {
-  const authorization = request.get('authorization');
-  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-    return authorization.substring(7);
-  }
-  return null;
-};
-
 // CREATE (POST) NEW ENTRY
 blogRouter.post('/', async (request, response, next) => {
   const body = await request.body;
 
-  const token = getTokenFrom(request);
-  const decodedToken = jwt.verify(token, process.env.SECRET);
-  if (!decodedToken.id) {
-    return response.status(401).json({ error: 'token missing or invalid' });
-  }
-
-  const user = await User.findById(decodedToken.id);
+  const user = request.user;
 
   const blog = new Blog({
     title: body.title,
@@ -52,8 +38,15 @@ blogRouter.post('/', async (request, response, next) => {
 
 // DELETE FROM DB
 blogRouter.delete('/:id', async (request, response, next) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response
+      .status(401)
+      .json({ error: 'token missing or invalid permissions' });
+  } else {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  }
 });
 
 // UPDATE ITEM ON DB
